requirements.txt
numpy~=1.18.4
tensorflow~=2.2.0
Pillow~=7.1.2